Norminette for glmath
