+++ glmath/srcs/camera.c
@@ -6,7 +6,7 @@
 /*   By: trossel <trossel@42lausanne.ch>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/03 07:47:50 by trossel           #+#    #+#             */
-/*   Updated: 2022/06/08 12:05:50 by trossel          ###   ########.fr       */
+/*   Updated: 2022/06/09 10:00:27 by trossel          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,12 +19,11 @@ t_ray	camera_init_ray(t_scene *s, int x, int y)
 
 	scale = tan(s->cam.fov / 2);
 	v3f_copy(&r.origin, &s->cam.pos);
-
-	r.direction.x = (2.0f * ((float)x + 0.5f) / (float)s->w - 1.0f) * s->ratio * scale;
+	r.direction.x = (2.0f * ((float)x + 0.5f) / (float)s->w - 1.0f);
+	r.direction.x *= s->ratio * scale;
 	r.direction.y = -(2.0f * ((float)y + 0.5f) / (float)s->h - 1.0f) * scale;
 	r.direction.z = 1.0f;
 	r.direction.w = 0.0f;
-
 	r.direction = cam2world(&s->cam, r.direction);
 	v3f_normalize(&r.direction);
 	return (r);
@@ -32,14 +31,13 @@ t_ray	camera_init_ray(t_scene *s, int x, int y)
 
 t_v3f	cam2world(t_camera *c, t_v3f p)
 {
-	t_v3f v;
+	t_v3f	v;
 
 	v.x = c->right.x * p.x + c->up.x * p.y + c->orien.x * p.z;
 	v.y = c->right.y * p.x + c->up.y * p.y + c->orien.y * p.z;
 	v.z = c->right.z * p.x + c->up.z * p.y + c->orien.z * p.z;
 	v.w = 0;
 	return (v);
-
 }
 
 /* The if is here because of gimbal lock */
@@ -54,13 +52,14 @@ void	camera_update_orien(t_camera *c, t_v3f orien)
 	}
 	else
 	{
-		c->right =  v3f_cross_product(&(t_v3f){0, 1.0f, 0, 0}, &c->orien);
+		c->right = v3f_cross_product(&(t_v3f){0, 1.0f, 0, 0}, &c->orien);
 		c->up = v3f_cross_product(&c->orien, &c->right);
 	}
 	v3f_normalize(&c->right);
 	v3f_normalize(&c->up);
 }
 
+/*
 t_v3f	look_at(const t_camera *cam, const t_scene *scene, int x, int y)
 {
 	t_v3f	left;
@@ -86,9 +85,12 @@ t_v3f	look_at(const t_camera *cam, const t_scene *scene, int x, int y)
 	cross = v3f_cross_product(&up, &vector);
 	v3f_normalize(&cross);
 
-	qr = qion_rotation_angle(-half_fov + ((y + ((float)(scene->w - scene->h) / 2)) * fov_ratio), &cross);
+	qr = qion_rotation_angle(
+			-half_fov + ((y + ((float)(scene->w - scene->h) / 2)) * fov_ratio),
+			&cross);
 	vector = qion_rotation(&vector, &qr);
 
 	v3f_normalize(&vector);
 	return (vector);
 }
+//*/
